pyproject.toml
[project]
name = "ai-engineering"
version = "0.1.0"
description = "Applied AI & ML — courses, projects, and experiments"
requires-python = ">= 3.12"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "ipykernel>=7.2",
    "python-dotenv>=1.0",
    "jupyter",
]

[project.optional-dependencies]
ml = [
    "scikit-learn",
    "torch",
]
llm = [
    "openai>=1.40",
    "gradio>=5.0",
]
data = [
    "simple-salesforce>=1.12",
    "databricks-sdk",
]
dev = [
    "pytest",
    "ruff",
]
all = [
    "ai-engineering[ml,llm,data,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
